# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/business-units": {
      "get": {
        "operationId": "BusinessUnitsController_list[0]_v1",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "business-units"
        ]
      },
      "post": {
        "operationId": "BusinessUnitsController_create[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusinessUnitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "business-units"
        ]
      }
    },
    "/business-units": {
      "get": {
        "operationId": "BusinessUnitsController_list[1]",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "business-units"
        ]
      },
      "post": {
        "operationId": "BusinessUnitsController_create[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusinessUnitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "business-units"
        ]
      }
    },
    "/v1/business-units/{buId}": {
      "patch": {
        "operationId": "BusinessUnitsController_update[0]_v1",
        "parameters": [
          {
            "name": "buId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessUnitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "business-units"
        ]
      }
    },
    "/business-units/{buId}": {
      "patch": {
        "operationId": "BusinessUnitsController_update[1]",
        "parameters": [
          {
            "name": "buId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessUnitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "business-units"
        ]
      }
    },
...
    "/v1/clients/{clientId}": {
      "get": {
        "operationId": "ClientsController_get[0]_v1",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "clients"
        ]
      },
      "patch": {
        "operationId": "ClientsController_update[0]_v1",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update client settings or move it through its lifecycle",
        "tags": [
          "clients"
        ]
      }
    },
    "/clients/{clientId}": {
      "get": {
        "operationId": "ClientsController_get[1]",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "clients"
        ]
      },
      "patch": {
        "operationId": "ClientsController_update[1]",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update client settings or move it through its lifecycle",
        "tags": [
          "clients"
        ]
      }
    },
...
    "/v1/authorization/check": {
      "post": {
        "operationId": "AuthorizationController_check[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizationCheckDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Server-side permission check for one user, client and permission (service token)",
        "tags": [
          "authorization"
        ]
      }
    },
    "/authorization/check": {
      "post": {
        "operationId": "AuthorizationController_check[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizationCheckDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Server-side permission check for one user, client and permission (service token)",
        "tags": [
          "authorization"
        ]
      }
    },
    "/v1/authorization/effective-permissions": {
      "post": {
        "operationId": "AuthorizationController_effectivePermissions[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EffectivePermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "All roles, modules and permissions a user holds for a client (service token)",
        "tags": [
          "authorization"
        ]
      }
    },
    "/authorization/effective-permissions": {
      "post": {
        "operationId": "AuthorizationController_effectivePermissions[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EffectivePermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "All roles, modules and permissions a user holds for a client (service token)",
        "tags": [
          "authorization"
        ]
      }
    },
    "/v1/internal/federation/clients/{clientId}": {
      "get": {
        "operationId": "FederationController_client[0]_v1",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Client configuration used by Identity Federation to validate embedded login",
        "tags": [
          "internal-federation"
        ]
      }
    },
    "/internal/federation/clients/{clientId}": {
      "get": {
        "operationId": "FederationController_client[1]",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Client configuration used by Identity Federation to validate embedded login",
        "tags": [
          "internal-federation"
        ]
      }
    },
    "/v1/internal/federation/users/{itsId}/assignments": {
      "get": {
        "operationId": "FederationController_assignments[0]_v1",
        "parameters": [
          {
            "name": "itsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Workspaces (role × scope) returned by POST /login",
        "tags": [
          "internal-federation"
        ]
      }
    },
    "/internal/federation/users/{itsId}/assignments": {
      "get": {
        "operationId": "FederationController_assignments[1]",
        "parameters": [
          {
            "name": "itsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Workspaces (role × scope) returned by POST /login",
        "tags": [
          "internal-federation"
        ]
      }
    },
    "/v1/internal/federation/assignments/resolve": {
      "post": {
        "operationId": "FederationController_resolve[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Validate a workspace for POST /select-scope; returns active scope + permission map",
        "tags": [
          "internal-federation"
        ]
      }
    },
    "/internal/federation/assignments/resolve": {
      "post": {
        "operationId": "FederationController_resolve[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Validate a workspace for POST /select-scope; returns active scope + permission map",
        "tags": [
          "internal-federation"
        ]
      }
    },
    "/v1/internal/federation/users/{itsId}/applications": {
      "get": {
        "operationId": "FederationController_applications[0]_v1",
        "parameters": [
          {
            "name": "itsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "required": true,
            "in": "query",
            "schema": {
              "example": "DEV",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Applications a user may launch from the Core Portal",
        "tags": [
          "internal-federation"
        ]
      }
    },
    "/internal/federation/users/{itsId}/applications": {
      "get": {
        "operationId": "FederationController_applications[1]",
        "parameters": [
          {
            "name": "itsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "required": true,
            "in": "query",
            "schema": {
              "example": "DEV",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Applications a user may launch from the Core Portal",
        "tags": [
          "internal-federation"
        ]
      }
    }
  },
  "info": {
    "title": "Miqaat Core Identity Authorization Service",
    "description": "Dynamic Business Unit / Utility / Application / Environment / Module / Role / Permission model, client registry and server-side authorization checks. No API keys: every call carries an RS256 bearer JWT verified via JWKS - service tokens (typ client-authentication+jwt) from registered service principals, or administrator access tokens (typ at+jwt) issued by Identity Federation. Never stores credentials.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT (RS256, verified via JWKS)",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateBusinessUnitDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "RMS"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ],
            "default": "active"
          },
          "tenant_id": {
            "type": "string",
            "description": "Only for ADMIN service principals; users always create in their own tenant"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateBusinessUnitDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          }
        }
      },
...
      "UpdateClientDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "SECURITY_REVIEW",
              "ACTIVE",
              "SUSPENDED",
              "RETIRED"
            ]
          },
          "reason": {
            "type": "string",
            "description": "Reason recorded in client status history"
          },
          "name": {
            "type": "string"
          },
          "authentication_mode": {
            "type": "string",
            "enum": [
              "EMBEDDED",
              "REDIRECT",
              "EMBEDDED_OR_REDIRECT"
            ]
          },
          "back_channel_logout_uri": {
            "type": "object",
            "nullable": true,
            "description": "null removes the back-channel logout URI"
          },
          "initiate_login_uri": {
            "type": "object",
            "nullable": true
          }
        }
      },
...
      "AuthorizationCheckDto": {
        "type": "object",
        "properties": {
          "its_id": {
            "type": "string",
            "example": "ITS12345"
          },
          "client_id": {
            "type": "string",
            "example": "rms-web-prod"
          },
          "module": {
            "type": "string",
            "example": "RMS_REGISTRATION"
          },
          "permission": {
            "type": "string",
            "example": "RMS_REGISTRATION_VIEW"
          }
        },
        "required": [
          "its_id",
          "client_id",
          "permission"
        ]
      },
      "CheckResponse": {
        "type": "object",
        "properties": {
          "allowed": {
            "type": "boolean"
          },
          "its_id": {
            "type": "string"
          },
          "client_id": {
            "type": "string"
          },
          "permission": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "example": "PERMISSION_DENIED"
          }
        },
        "required": [
          "allowed"
        ]
      },
      "EffectivePermissionsDto": {
        "type": "object",
        "properties": {
          "its_id": {
            "type": "string",
            "example": "ITS12345"
          },
          "client_id": {
            "type": "string",
            "example": "rms-web-prod"
          }
        },
        "required": [
          "its_id",
          "client_id"
        ]
      },
      "ResolveAssignmentDto": {
        "type": "object",
        "properties": {
          "its_id": {
            "type": "string",
            "example": "31267890"
          },
          "role_id": {
            "type": "string"
          },
          "scope_type": {
            "type": "string",
            "enum": [
              "CORE",
              "BUSINESS_UNIT",
              "UTILITY"
            ]
          },
          "scope_id": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "its_id",
          "role_id",
          "scope_type"
        ]
      }
    }
  }
}